Also update cc_toolchain.toolchain_identifier when CC_TOOLCHAIN_NAME is set
Since we specify toolchain_identifier in cc_toolchain now, it is much more
strict in selecting a CROSSTOOL toolchain. This broke when CC_TOOLCHAIN_NAME
environment variable was set, as that only renamed toolchain identifier in the
CROSSTOOL, but not in the cc_toolchain. This cl fixes that.

This fixed https://github.com/bazelbuild/bazel-toolchains/issues/198.

RELNOTES: None
PiperOrigin-RevId: 217666695